Fix fetcher
The brainz_fetcher, which is currently the only fetcher that is started,
could not run. The reason was using `Rails.application.config_for`
in the ImportConnection module.

I resolved the problem by loading config/application and requiring
all files under app/models in `config/containers/import_queue/queue_cmd.rb`.

Another solution might be to use a rake-task to start a fetcher.
This should set the environment correctly.