Switch to RSpec feature specs
Maybe a slightly chaotic commit!

While working on the order_import.feature I decided that it might be
a good idea to switch to RSpec feature specs instead of using cucumber.

The reason was the missing support for devise user sign in and sign out.
Another reason was the dull feeling that cucumber is no longer
supported well.