Set the default for #state in Importable
The class ImportOrder does not include BufferableImport.
The default for #state ("open") was set in BufferableImport.
This means for an ImportOrder there was no default for #state.
This caused an error unless the user sets a value for #state.
So the setting of the default for #state was moved from
BufferableImport to Importable.